agri-chatbot-backend/requirements.txt
fastapi
uvicorn
python-dotenv
ollama
deep-translator
langchain-community
faiss-cpu
sentence-transformers
